requirements.txt
Django>=4.2,<5.0
django-guardian>=2.4.0
reportlab>=4.0
Pillow>=10.0
django-crispy-forms>=2.0
crispy-bootstrap5>=0.7
django-widget-tweaks>=1.5
djangorestframework>=3.14
django-filter>=23.0,<25.0
qrcode[pil]>=7.4
openpyxl>=3.1
